[perl] Name explicit List::Util version dependency
Although List::Util is a core module (better: it's dual life), it only grew support for 'any' and 'all'
in 1.33, which doesn't come with some of the older Perl versions. Specifying the number explicitly
makes older Perls install this module from CPAN.

No need to update the CHANGELOG for this: it's already mentioned, this commit just reverts the
Renovate change 1.33->1.62 and adds the reasoning behind the version number to the place where
others are most likely to find it: the cpanfile itself.